requirements.txt
beautifulsoup4==4.7.1
requests==2.17.3
mecab-python3==0.996.1
